[MNG-8564] Move to maven-compat Plexus legacy
This merely moves a newly introduced legacy support class to
maven-compat.

---

https://issues.apache.org/jira/browse/MNG-8564
